Refactor SearchModule, including: 1. Moving functions that are related to 'search' to their correct modules. 2. Creating 'SearchPageHelper' to help process Search page Presentation layer.
