+++ Source/Plugins/Core/com.equella.core/js/tsrc/modules/CollectionsModule.ts
@@ -20,6 +20,20 @@ import { API_BASE_URL } from "../AppConfig";
 import { listEntities } from "./OEQHelpers";
 
 /**
+ * A simplified type of Collection used when only collection uuid and name are required.
+ */
+export interface Collection {
+  /**
+   * Collection's uuid.
+   */
+  uuid: string;
+  /**
+   * Collection's name.
+   */
+  name: string;
+}
+
+/**
  * Provides a simple Map<string,string> summary of available collections, where K is the UUID
  * and V is the collections's name.
  *
@@ -37,16 +51,20 @@ export const collectionListSummary = (
     (results) => results
   );
 };
+
 /**
- * A simplified type of Collection used when only collection uuid and name are required.
+ * Find a list of Collections by ID.
+ *
+ * @param collectionUuids Collection UUIDs used to filter the list of all Collections.
  */
-export interface Collection {
-  /**
-   * Collection's uuid.
-   */
-  uuid: string;
-  /**
-   * Collection's name.
-   */
-  name: string;
-}
+export const findCollectionsByUuid = async (
+  collectionUuids: string[]
+): Promise<Collection[] | undefined> => {
+  const collectionList = await collectionListSummary([
+    OEQ.Acl.ACL_SEARCH_COLLECTION,
+  ]);
+  const filteredCollectionList = collectionList.filter((c) =>
+    collectionUuids.includes(c.uuid)
+  );
+  return filteredCollectionList.length > 0 ? filteredCollectionList : undefined;
+};
